test(create-plugin): exit-code should be non-zero when an error occurs
